Added pinterest widget, made carousel fixes (work in progress)
Added pinterest widget, added styles to center it. Need to work on more
styles, responsive, width, etc.
Tweaked carousel styles to fix found issues. Had to add the hovering
text to each individual slide to allow for slide animation and to allow
for next prev buttons to appear on hover.